Fix package name "android"
